# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List accessible tenants
      responses:
        "200":
          description: Accessible tenants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantSummary" }
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create and activate a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantInput" }
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantSummary" }
  /tenants/{id}/activate:
    post:
      operationId: activateTenant
      tags: [tenants]
      summary: Activate an accessible tenant
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Active tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantSummary" }
  /tenants/{id}/members:
    get:
      operationId: listTenantMembers
      tags: [tenants]
      summary: List tenant members
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tenant members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantMember" }
    post:
      operationId: addTenantMember
      tags: [tenants]
      summary: Add or update an existing local user
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantMemberInput" }
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantMember" }
  /tenants/{id}/members/{userId}:
    delete:
      operationId: deleteTenantMember
      tags: [tenants]
      summary: Remove a tenant member
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List configured ComfyUI servers
      responses:
        "200":
          description: Server cards without private connection details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComfyServer"
    post:
      operationId: createServer
      tags: [servers]
      summary: Add a ComfyUI server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateInput"
      responses:
        "201":
          description: Created server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComfyServer"
  /servers/{id}:
    patch:
      operationId: updateServer
      tags: [servers]
      summary: Update a ComfyUI server
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdateInput"
      responses:
        "200":
          description: Updated server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComfyServer"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Remove a server
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /servers/{id}/configuration:
    get:
      operationId: getServerConfiguration
      tags: [servers]
      summary: Read saved ComfyUI connection details for editing
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Editable ComfyUI connection configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerConnectionConfiguration"
  /servers/{id}/test:
    post:
      operationId: testServerConnection
      tags: [servers]
      summary: Test a configured server connection
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Connection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTest"
  /servers/{id}/queue:
    get:
      operationId: getServerQueue
      tags: [servers]
      summary: Read the queue for a configured server
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Queue snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueSnapshot"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    TenantRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER]
...
    TenantSummary:
      type: object
      required: [id, name, slug, role]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        role: { $ref: "#/components/schemas/TenantRole" }
...
    TenantInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    TenantMember:
      type: object
      required: [userId, email, displayName, role]
      properties:
        userId: { type: string }
        email: { type: ["string", "null"] }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/TenantRole" }
    TenantMemberInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { $ref: "#/components/schemas/TenantRole" }
...
    ComfyServer:
      type: object
      required: [id, displayName, hostname, gpuName, vramGb, tags, enabled, priority, status, queueSize, activeJobCount, lastHeartbeat, supportedWorkflowCount]
      properties:
        id: { type: string }
        displayName: { type: string }
        hostname: { type: string }
        gpuName: { type: ["string", "null"] }
        vramGb: { type: ["number", "null"] }
        tags: { type: array, items: { type: string } }
        enabled: { type: boolean }
        priority: { type: number }
        status: { type: string, enum: [ONLINE, OFFLINE, UNKNOWN] }
        queueSize: { type: number }
        activeJobCount: { type: number }
        lastHeartbeat: { type: ["string", "null"] }
        supportedWorkflowCount: { type: number }
        memoryUsedGb: { type: ["number", "null"] }
    ServerCreateInput:
      type: object
      required: [displayName, apiBaseUrl, websocketUrl, priority]
      properties:
        displayName: { type: string, minLength: 1 }
        apiBaseUrl: { type: string }
        websocketUrl: { type: string }
        gpuName: { type: ["string", "null"] }
        vramGb: { type: ["number", "null"] }
        tags: { type: array, items: { type: string } }
        enabled: { type: boolean, default: true }
        priority: { type: number, minimum: 0 }
        maxConcurrentJobs: { type: ["number", "null"], minimum: 1 }
    ServerUpdateInput:
      type: object
      properties:
        displayName: { type: string, minLength: 1 }
        apiBaseUrl: { type: string }
        websocketUrl: { type: string }
        gpuName: { type: ["string", "null"] }
        vramGb: { type: ["number", "null"] }
        tags: { type: array, items: { type: string } }
        enabled: { type: boolean }
        priority: { type: number, minimum: 0 }
        maxConcurrentJobs: { type: ["number", "null"], minimum: 1 }
    ServerConnectionConfiguration:
      type: object
      required: [id, apiBaseUrl, websocketUrl]
      properties:
        id: { type: string }
        apiBaseUrl: { type: string }
        websocketUrl: { type: string }
    ConnectionTest:
      type: object
      required: [connected, message, server, gpu, vramGb]
      properties:
        connected: { type: boolean }
        message: { type: string }
        server: { type: ["string", "null"] }
        gpu: { type: ["string", "null"] }
        vramGb: { type: ["number", "null"] }
    QueueSnapshot:
      type: object
      required: [queueRunning, queuePending]
      properties:
        queueRunning: { type: number }
        queuePending: { type: number }
        queueRunningIds: { type: array, items: { type: string } }
        queuePendingIds: { type: array, items: { type: string } }
...