# lib/api-zod/src/generated/types/tenantInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TenantInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}