# lib/api-zod/src/generated/types/tenantMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TenantRole } from './tenantRole';

export interface TenantMember {
  userId: string;
  /** @nullable */
  email: string | null;
  displayName: string;
  role: TenantRole;
}